Workaround deprecated 1.2 API
The DBC 1.2 REST API has been deprecated and some of the functionality
stopped working. Some of the endpoints that used to return 4xx
unauthorized responses, now may return 500. Unfortunately, that means
that the Apache HTTP Client does not attempt to authenticate using the
Authorization header (basic HTTP auth), because by design it does not
pass that header unless challenged. To fix that, we pass the context
around that contains a credentials cache.